Assignment 2/pyproject.toml
[project]
name = "assignment-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
